Make relying party data configurable
This helps when using the same key on multiple wikis.

It can only be used if all the wikis use the same root domain, by
specifying the relying partyID to the common part of the comain (must
include a period ), eg. wiki1.mywiki.com and wiki2.mywiki.com is ok,
but wiki1.com and wiki2.com are not. Also, relying party name must be
the same for all wikis involved.

These limitations are built in webauthn and cannot be bypassed.

Bug: T244088
Change-Id: I9878af25294ec4fea7126f6b7f96e0014c77147e