Auth: Add authentication request indicating used module
* Add TwoFactorModuleSelectAuthenticationRequest class which holds
  information about which 2FA module is used during authentication
* When the user has multiple 2FA methods enabled, use the one
  designated by TwoFactorModuleSelectAuthenticationRequest if
  that's included in the request data.
* On the login page, it is represented as a dropdown for selecting
  a different 2FA module.

Also move some integration tests to the integration test dir.

From the user's perspective, this patch is a no-op since having
multiple authentication methods isn't allowed at the moment
(OATHUserRepository::createKey() will throw when trying to set up
a second method).

The goal of the patch is to enable developers to use multiple
authenticators, not to provide a sensible UI. There are many
questionable aspects:
* The dropdown is way too conspicuous, normally there'd be a small
  link or such for picking a different method.
* If the user selects a different method, the form elements for
  the old method should be removed - e.g. the only way to switch
  away from TOTP is to first type something random in the TOTP
  code field as it has the 'reqired' HTML attribute.
* It also doesn't work well with the WebAuthn form's auto-reload
  on error.

Since the UI is likely to undergo major changes, it didn't seem
worth spending time on fixing those issues for now.

Bug: T242031
Change-Id: I8df6e12d6cd355f3e2db4e68078a0e65a2c2268e