Use paragraphs and avoid wrapping with labels
Use paragraphs instead of linebreaks to create the text blocks.
This automatically provides margins for surrounding content like the
recovery code list and its buttons and makes the text more readable.

The info-type section by default will wrap everything in a label. Sometimes
this is desirable, but not so much in this layout, which is more a
continuation of the previous field.
The rawrow option allows us to disable this label wrapping.

Change-Id: I3d1363e3ac3d7d27a9c4d2d4b121edf525ddc513