Don't require separate authentication to disable a module
In order to allow the user to disable a module, we were requiring them
to authenticate with that module. To disable TOTP, you had to put in a
TOTP code, and to disable WebAuthn you had to authenticate with one of
your WebAuthn keys. But this is undesirable because it doesn't allow a
user who has lost their TOTP to disable it by authenticating with
WebAuthn or vice versa.

Now that all views and actions on Special:OATHManage require
reauthentication, this extra authentication step is unnecessary, because
the user already had to authenticate with one of their 2FA methods
during reauthentication. This change removes it, merging the disable
warning and the disable form into one step.

- Remove per-method disable forms
- Remove the "disable warning" triggered by the &warn=1 parameter
- Add a module-agnostic disable form, which contains the disable warning
  text instead of an authentication challenge
- Remove now-unused i18n messages

Bug: T402989
Change-Id: I8334a8572a61e65590dd0a2ba9c49c95abfdca67