Provide backward compatibilty for the mobile apps' login flow
The Wikipedia mobile apps for Android and iOS use the action=clientlogin
API, and don't support WebAuthn. To avoid breaking them, we have to
ensure that TOTP is offered first. If WebAuthn were to be offered first,
the "requests" array they would receive would contain only a
WebAuthnAuthenticationRequest and a
TwoFactorModuleSelectAuthenticationRequest, neither of which is
currently recognized by the apps.

We should be able to remove this hack soonish, once the apps have been
updated to handle the scenario where WebAuthn is presented first and
they need to switch to TOTP.

Bug: T399654
Change-Id: I6607e4e54e76e7a7ca315ece013875b8d07e42bc