Support more attestation formats
Add all supported attestation formats[1], except SafetyNet which
is deprecated[2].

I don't think this does much in practice, missing attestation support
results in a "The attestation statement format ... is not supported"
error being logged, and we don't have any of those. (Probably because
we set the attestation conveyance preference[3] to 'none' in
Authenticator::getRegisterInfo()?) But it still makes more sense than
adding some but not all formats.

[1] https://www.w3.org/TR/webauthn-2/#sctn-defined-attestation-formats
[2] https://developer.android.com/privacy-and-security/safetynet/deprecation-timeline
[3] https://www.w3.org/TR/webauthn-2/#attestation-conveyance

Change-Id: Id23b53ed3d1cbcdb7345695023f0cbcd8b329f1c