OATHManage: Fix cancel button form action
We were setting the action for the form that wraps the cancel button to
the URL of the special page. That works if the URL is something like
`/wiki/Special:AccountSecurity`, but not if it's
`/w/index.php?title=Special:AccountSecurity`. In the latter case, the
query string is stripped, and we end up sending the user to the main
page.

Instead, set the form action to index.php (as we do for all the other
forms on this page) and add a hidden `title` field.

Bug: T405030
Change-Id: If74f4aa2721192a36a83b92d6093b57ba59a9941