RecoveryCodesTrait: Add issuer to the download filename, make translateable
I don't know about you, but I have recovery/backup/scratch codes for
a dozen sites. The recovery mechanism will be easier to use if the
users don't have to rename the file to know which site the recovery
codes belong to, and it lets us document the expected filename
somewhere if most users are likely to have the same one.

Instead of 'recovery-codes.txt', this will use something like
'Wikimedia - recovery codes.txt', depending on the site config.

The "recovery codes" part is now a translateable message. The rest
of the string is not translateable, so that translators can't
accidentally break it, and so that the format is consistent
regardless of the language. Characters other than letters and some
punctuation (arbitrarily chosen) are removed from the filename.

We could consider also including the username and the date somewhere.
Some other systems with recovery codes do that, although they are
usually included in the file content rather than filename.

Bug: T406382
Change-Id: Idacd91d8bd59b1ecb98e66dcc7da26c7164baa9a