OATHManage: Don't always set the page title to "Create new recovery codes"
Counter-intuitively, the getDescriptors() method of each form is always
run on every page view of Special:AccountSecurity, so setting the page
title there set it for every page view. Instead, move this to getHTML(),
which only runs when the form is actually displayed.

In a future change I'll restructure this so that we don't construct
every form on every page view.

Change-Id: I79c77d592f334b5f4acf94d1cb978617e11e051b